Add free functions for bonus data handling
