+++ src/mlx/hooks.c
@@ -6,7 +6,7 @@
 /*   By: mdekker <mdekker@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2024/01/17 16:24:40 by mdekker       #+#    #+#                 */
-/*   Updated: 2024/03/19 15:29:41 by mdekker       ########   odam.nl         */
+/*   Updated: 2024/03/19 15:56:45 by mdekker       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,11 @@
 
 static mouse_mode_t	mode_set(t_data *data, mouse_mode_t mode)
 {
-	static mouse_mode_t	current_mode;
+	static mouse_mode_t	current_mode = MLX_MOUSE_DISABLED;
+	static double		last_press = 0;
 
-	current_mode = MLX_MOUSE_DISABLED;
+	if (mlx_get_time() - last_press < 0.5)
+		return (current_mode);
 	if (mode == MLX_MOUSE_DISABLED)
 	{
 		current_mode = MLX_MOUSE_DISABLED;
@@ -53,9 +55,8 @@ static void	toggle_all_doors(t_map *map)
 {
 	int				i;
 	int				j;
-	static double	last_press;
+	static double	last_press = 0;
 
-	last_press = 0;
 	i = 0;
 	if (mlx_get_time() - last_press < 0.5)
 		return ;
@@ -77,9 +78,8 @@ static void	toggle_all_doors(t_map *map)
 
 static void	esq_hook(t_data *data)
 {
-	static double	last_press;
+	static double	last_press = 0;
 
-	last_press = 0;
 	if (mlx_is_key_down(data->mlx, MLX_KEY_LEFT_SHIFT))
 	{
 		if (mlx_get_time() - last_press < 0.5)
@@ -117,6 +117,5 @@ void	key_hook(void *param)
 	{
 		data->textures.sprite.current = (data->textures.sprite.current + 1)
 			% data->textures.sprite.images.length;
-		// draw_minimap(data);m,÷:"[p
 	}
 }
